+++ tests/Main.hs
@@ -8,10 +8,11 @@ module Main (main) where
 import Prelude ()
 import Prelude.Compat
 
-import Test.Tasty
+import Test.Hspec
 
 import qualified TestCompiler
 import qualified TestCoreFn
+import qualified TestCst
 import qualified TestDocs
 import qualified TestHierarchy
 import qualified TestPrimDocs
@@ -33,34 +34,19 @@ main = do
   heading "Updating support code"
   TestUtils.updateSupportCode
 
-  ideTests <- TestIde.main
-  compilerTests <- TestCompiler.main
-  makeTests <- TestMake.main
-  psciTests <- TestPsci.main
-  pscBundleTests <- TestBundle.main
-  coreFnTests <- TestCoreFn.main
-  docsTests <- TestDocs.main
-  primDocsTests <- TestPrimDocs.main
-  publishTests <- TestPscPublish.main
-  hierarchyTests <- TestHierarchy.main
-  graphTests <- TestGraph.main
-
-  defaultMain $
-    testGroup
-      "Tests"
-      [ compilerTests
-      , makeTests
-      , psciTests
-      , pscBundleTests
-      , ideTests
-      , coreFnTests
-      , docsTests
-      , primDocsTests
-      , publishTests
-      , hierarchyTests
-      , graphTests
-      ]
-
+  hspec $ do
+    describe "cst" TestCst.spec
+    describe "ide" TestIde.spec
+    describe "compiler" TestCompiler.spec
+    describe "make" TestMake.spec
+    describe "psci" TestPsci.spec
+    describe "bundle" TestBundle.spec
+    describe "corefn" TestCoreFn.spec
+    describe "docs" TestDocs.spec
+    describe "prim-docs" TestPrimDocs.spec
+    describe "publish" TestPscPublish.spec
+    describe "hierarchy" TestHierarchy.spec
+    describe "graph" TestGraph.spec
   where
   heading msg = do
     putStrLn ""